code/aml-embeddings-fabric-unstructured-data-flow/requirements.txt
azure-ai-ml
azureml-fsspec
azure-search-documents
mltable
keyrings.alt